requirements.txt

sloth-framework
